fix id to BigDecimal in BuildingRepo

+++ Core-CRM/src/main/java/com/pm/corecrm/repository/TaskRepository.java
@@ -6,6 +6,7 @@ import com.pm.corecrm.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -16,5 +17,5 @@ public interface TaskRepository extends JpaRepository<Task,Long> {
     // Поиск по статусу
     List<Task> findByStatus(TaskStatus status);
     // Поиск задач по ID здания (Building)
-    List<Task> findByBuildingId(Long buildingId);
+    List<Task> findByBuildingId(BigDecimal buildingId);
 }
